Offline login
A lot of commits were squashed: these were the messages

Adjust for partial login state on the worker side, #3888


Handle partial and full login on the client side, #3888


Attempt to log in when coming online if async login fails, #3888


Allow only Premium customers to log in offline, #3888

We return result from resumeLogin to not add a bespoke error class for
our use case. We cannot use UserError there either. We should probably
convert more expected errors to results as it's hard to know what
should be handled when we reach UI layer.


Allow fixing credentials after offline login, #3888

Also improve progress display in password prompt dialog


Allow re-initializing cache storage when the first login fails, #3888


Improve LoginFacadeTest

Do not keep credentials for retry if they cannot be used again, #3888


Fix initialising indexer from LoginFacade, #3888


Add offlineLoginRequiresPremium_msg for de, de_sie, #3888

Cleanup facades fields after review with jom


Cleanup UserFacade, fix tests


Fix waiting for full login in LoginController


Split AsyncActions into partial and full login


Cleanup registering of post-login actions


Change offlineDb parameter on CredentialsProvider after review with jom


Remove out-of-sync handling on cache init

Current approach does not work in offline, we are going to reimplement
it.

See #4067

Suppress connection error during offline login, #3888

Do not wait for ConnectionError in LoginFacadeTest

* ConnectionError is suppressed in offline login

Prevent endless waiting for contacts after offline login, fix #4078

* When the user is logged in offline the indexer has not been initialized
 and searching for contacts will never resolve.

Handle UserGroupKeyNotFoundError, see #4078

The userGroupKey is saved after login. However, when the user logs in
offline, the key is not saved and cannot be used to encrypt. We handle
this in the same way we handle ConnectionErrors.

+++ src/misc/credentials/CredentialsProvider.ts
@@ -133,8 +133,9 @@ export interface ICredentialsProvider {
 	/**
 	 * Deletes stored credentials with specified userId.
 	 * No-op if credentials are not there.
+	 * @param opts.deleteOfflineDb whether to delete offline database. Will delete by default.
 	 */
-	deleteByUserId(userId: Id): Promise<void>
+	deleteByUserId(userId: Id, opts?: {deleteOfflineDb: boolean}): Promise<void>
 
 	/**
 	 * Sets the credentials encryption mode, i.e. how the intermediate key used for encrypting credentials is protected.
@@ -209,8 +210,6 @@ export class CredentialsProvider implements ICredentialsProvider {
 			decrypted.databaseKey = await this.databaseKeyFactory.generateKey()
 
 			if (decrypted.databaseKey != null) {
-				// TODO this might prompt the user to unlock the keychain again
-				// 		We should figure out what to do about this
 				const reEncrypted = await this.credentialsEncryption.encrypt(decrypted)
 				this.storage.store(reEncrypted)
 			}
@@ -226,13 +225,15 @@ export class CredentialsProvider implements ICredentialsProvider {
 		})
 	}
 
-	async deleteByUserId(userId: Id): Promise<void> {
+	async deleteByUserId(userId: Id, opts: {deleteOfflineDb: boolean} = {deleteOfflineDb: true}): Promise<void> {
 		const event: CredentialsDeletedEvent = {
 			name: CREDENTIALS_DELETED_EVENT,
 			userId,
 		}
 		await this.interWindowEventBus?.send(event)
-		await this.offlineDbFacade?.deleteDatabaseForUser(userId)
+		if (opts?.deleteOfflineDb) {
+			await this.offlineDbFacade?.deleteDatabaseForUser(userId)
+		}
 		this.storage.deleteByUserId(userId)
 	}
